Network.AWS.ELBv2.CreateRule

Description

Creates a rule for the specified listener.

Each rule can have one action and one condition. Rules are evaluated in priority order, from the lowest value to the highest value. When the condition for a rule is met, the specified action is taken. If no conditions are met, the default action for the default rule is taken. For more information, see Listener Rules in the Application Load Balancers Guide .

To view your current rules, use DescribeRules . To update a rule, use ModifyRule . To set the priorities of your rules, use SetRulePriorities . To delete a rule, use DeleteRule .

Creating a Request

createRule

Arguments

:: Text
crListenerARN

-> Natural
crPriority

-> CreateRule

Creates a value of CreateRule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crListenerARN - The Amazon Resource Name (ARN) of the listener.

crConditions - A condition. Each condition specifies a field name and a single value. If the field name is host-header , you can specify a single host name (for example, my.example.com). A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * - . * * (matches 0 or more characters) * ? (matches exactly 1 character) If the field name is path-pattern , you can specify a single path pattern. A path pattern is case sensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * _ - . $ / ~ " ' @ : + * & (using &amp;) * * (matches 0 or more characters) * ? (matches exactly 1 character)

crPriority - The priority for the rule. A listener can't have multiple rules with the same priority.

crActions - An action. Each action has the type forward and specifies a target group.

data CreateRule

See: createRule smart constructor.

Request Lenses

crListenerARN :: Lens' CreateRule Text

The Amazon Resource Name (ARN) of the listener.

crConditions :: Lens' CreateRule [RuleCondition]

A condition. Each condition specifies a field name and a single value. If the field name is host-header , you can specify a single host name (for example, my.example.com). A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * - . * * (matches 0 or more characters) * ? (matches exactly 1 character) If the field name is path-pattern , you can specify a single path pattern. A path pattern is case sensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * _ - . $ / ~ " ' @ : + * & (using &amp;) * * (matches 0 or more characters) * ? (matches exactly 1 character)

crPriority :: Lens' CreateRule Natural

The priority for the rule. A listener can't have multiple rules with the same priority.

crActions :: Lens' CreateRule [Action]

An action. Each action has the type forward and specifies a target group.

Destructuring the Response

createRuleResponse

Arguments

:: Int
crrsResponseStatus

-> CreateRuleResponse

Creates a value of CreateRuleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsRules - Information about the rule.

crrsResponseStatus - -- | The response status code.

data CreateRuleResponse

See: createRuleResponse smart constructor.

Response Lenses

crrsRules :: Lens' CreateRuleResponse [Rule]

Information about the rule.

crrsResponseStatus :: Lens' CreateRuleResponse Int

- | The response status code.