Network.AWS.ELBv2.CreateTargetGroup

Description

Creates a target group.

To register targets with the target group, use RegisterTargets . To update the health check settings for the target group, use ModifyTargetGroup . To monitor the health of targets in the target group, use DescribeTargetHealth .

To route traffic to the targets in a target group, specify the target group in an action using CreateListener or CreateRule .

To delete a target group, use DeleteTargetGroup .

For more information, see Target Groups for Your Application Load Balancers in the Application Load Balancers Guide .

Creating a Request

createTargetGroup

Arguments

:: Text
ctgName

-> ProtocolEnum
ctgProtocol

-> Natural
ctgPort

-> Text
ctgVPCId

-> CreateTargetGroup

Creates a value of CreateTargetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctgMatcher - The HTTP codes to use when checking for a successful response from a target. The default is 200.

ctgHealthCheckPath - The ping path that is the destination on the targets for health checks. The default is /.

ctgUnhealthyThresholdCount - The number of consecutive health check failures required before considering a target unhealthy. The default is 2.

ctgHealthCheckIntervalSeconds - The approximate amount of time, in seconds, between health checks of an individual target. The default is 30 seconds.

ctgHealthyThresholdCount - The number of consecutive health checks successes required before considering an unhealthy target healthy. The default is 5.

ctgHealthCheckProtocol - The protocol the load balancer uses when performing health checks on targets. The default is the HTTP protocol.

ctgHealthCheckTimeoutSeconds - The amount of time, in seconds, during which no response from a target means a failed health check. The default is 5 seconds.

ctgHealthCheckPort - The port the load balancer uses when performing health checks on targets. The default is traffic-port , which indicates the port on which each target receives traffic from the load balancer.

ctgName - The name of the target group. This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

ctgProtocol - The protocol to use for routing traffic to the targets.

ctgPort - The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target.

ctgVPCId - The identifier of the virtual private cloud (VPC).

data CreateTargetGroup

See: createTargetGroup smart constructor.

Request Lenses

ctgMatcher :: Lens' CreateTargetGroup (Maybe Matcher)

The HTTP codes to use when checking for a successful response from a target. The default is 200.

ctgHealthCheckPath :: Lens' CreateTargetGroup (Maybe Text)

The ping path that is the destination on the targets for health checks. The default is /.

ctgUnhealthyThresholdCount :: Lens' CreateTargetGroup (Maybe Natural)

The number of consecutive health check failures required before considering a target unhealthy. The default is 2.

ctgHealthCheckIntervalSeconds :: Lens' CreateTargetGroup (Maybe Natural)

The approximate amount of time, in seconds, between health checks of an individual target. The default is 30 seconds.

ctgHealthyThresholdCount :: Lens' CreateTargetGroup (Maybe Natural)

The number of consecutive health checks successes required before considering an unhealthy target healthy. The default is 5.

ctgHealthCheckProtocol :: Lens' CreateTargetGroup (Maybe ProtocolEnum)

The protocol the load balancer uses when performing health checks on targets. The default is the HTTP protocol.

ctgHealthCheckTimeoutSeconds :: Lens' CreateTargetGroup (Maybe Natural)

The amount of time, in seconds, during which no response from a target means a failed health check. The default is 5 seconds.

ctgHealthCheckPort :: Lens' CreateTargetGroup (Maybe Text)

The port the load balancer uses when performing health checks on targets. The default is traffic-port , which indicates the port on which each target receives traffic from the load balancer.

ctgName :: Lens' CreateTargetGroup Text

The name of the target group. This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

ctgProtocol :: Lens' CreateTargetGroup ProtocolEnum

The protocol to use for routing traffic to the targets.

ctgPort :: Lens' CreateTargetGroup Natural

The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target.

ctgVPCId :: Lens' CreateTargetGroup Text

The identifier of the virtual private cloud (VPC).

Destructuring the Response

createTargetGroupResponse

Arguments

:: Int
ctgrsResponseStatus

-> CreateTargetGroupResponse

Creates a value of CreateTargetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctgrsTargetGroups - Information about the target group.

ctgrsResponseStatus - -- | The response status code.

data CreateTargetGroupResponse

See: createTargetGroupResponse smart constructor.

Response Lenses

ctgrsTargetGroups :: Lens' CreateTargetGroupResponse [TargetGroup]

Information about the target group.

ctgrsResponseStatus :: Lens' CreateTargetGroupResponse Int

- | The response status code.