Network.AWS.ELBv2.DeleteListener

Description

Deletes the specified listener.

Alternatively, your listener is deleted when you delete the load balancer it is attached to using DeleteLoadBalancer .

Creating a Request

deleteListener

Arguments

:: Text
dlListenerARN

-> DeleteListener

Creates a value of DeleteListener with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlListenerARN - The Amazon Resource Name (ARN) of the listener.

data DeleteListener

See: deleteListener smart constructor.

Request Lenses

dlListenerARN :: Lens' DeleteListener Text

The Amazon Resource Name (ARN) of the listener.

Destructuring the Response

deleteListenerResponse

Arguments

:: Int
dlrsResponseStatus

-> DeleteListenerResponse

Creates a value of DeleteListenerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlrsResponseStatus - -- | The response status code.

data DeleteListenerResponse

See: deleteListenerResponse smart constructor.

Response Lenses

dlrsResponseStatus :: Lens' DeleteListenerResponse Int

- | The response status code.