Network.AWS.ELBv2.DescribeSSLPolicies

Description

Describes the specified policies or all policies used for SSL negotiation.

For more information, see Security Policies in the Application Load Balancers Guide .

Creating a Request

describeSSLPolicies :: DescribeSSLPolicies

Creates a value of DescribeSSLPolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dspNames - The names of the policies.

dspMarker - The marker for the next set of results. (You received this marker from a previous call.)

dspPageSize - The maximum number of results to return with this call.

data DescribeSSLPolicies

See: describeSSLPolicies smart constructor.

Request Lenses

dspNames :: Lens' DescribeSSLPolicies [Text]

The names of the policies.

dspMarker :: Lens' DescribeSSLPolicies (Maybe Text)

The marker for the next set of results. (You received this marker from a previous call.)

dspPageSize :: Lens' DescribeSSLPolicies (Maybe Natural)

The maximum number of results to return with this call.

Destructuring the Response

describeSSLPoliciesResponse

Arguments

:: Int
dsprsResponseStatus

-> DescribeSSLPoliciesResponse

Creates a value of DescribeSSLPoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsprsSSLPolicies - Information about the policies.

dsprsNextMarker - The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dsprsResponseStatus - -- | The response status code.

data DescribeSSLPoliciesResponse

See: describeSSLPoliciesResponse smart constructor.

Response Lenses

dsprsSSLPolicies :: Lens' DescribeSSLPoliciesResponse [SSLPolicy]

Information about the policies.

dsprsNextMarker :: Lens' DescribeSSLPoliciesResponse (Maybe Text)

The marker to use when requesting the next set of results. If there are no additional results, the string is empty.

dsprsResponseStatus :: Lens' DescribeSSLPoliciesResponse Int

- | The response status code.