Network.AWS.ELBv2.ModifyListener

Description

Modifies the specified properties of the specified listener.

Any properties that you do not specify retain their current values. However, changing the protocol from HTTPS to HTTP removes the security policy and SSL certificate properties. If you change the protocol from HTTP to HTTPS, you must add the security policy and server certificate.

Creating a Request

modifyListener

Arguments

:: Text
mlListenerARN

-> ModifyListener

Creates a value of ModifyListener with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlSSLPolicy - The security policy that defines which protocols and ciphers are supported. For more information, see Security Policies in the Application Load Balancers Guide .

mlProtocol - The protocol for connections from clients to the load balancer.

mlDefaultActions - The default actions.

mlCertificates - The SSL server certificate.

mlPort - The port for connections from clients to the load balancer.

mlListenerARN - The Amazon Resource Name (ARN) of the listener.

data ModifyListener

See: modifyListener smart constructor.

Request Lenses

mlSSLPolicy :: Lens' ModifyListener (Maybe Text)

The security policy that defines which protocols and ciphers are supported. For more information, see Security Policies in the Application Load Balancers Guide .

mlProtocol :: Lens' ModifyListener (Maybe ProtocolEnum)

The protocol for connections from clients to the load balancer.

mlDefaultActions :: Lens' ModifyListener [Action]

The default actions.

mlCertificates :: Lens' ModifyListener [Certificate]

The SSL server certificate.

mlPort :: Lens' ModifyListener (Maybe Natural)

The port for connections from clients to the load balancer.

mlListenerARN :: Lens' ModifyListener Text

The Amazon Resource Name (ARN) of the listener.

Destructuring the Response

modifyListenerResponse

Arguments

:: Int
mlrsResponseStatus

-> ModifyListenerResponse

Creates a value of ModifyListenerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlrsListeners - Information about the modified listeners.

mlrsResponseStatus - -- | The response status code.

data ModifyListenerResponse

See: modifyListenerResponse smart constructor.

Response Lenses

mlrsListeners :: Lens' ModifyListenerResponse [Listener]

Information about the modified listeners.

mlrsResponseStatus :: Lens' ModifyListenerResponse Int

- | The response status code.