Network.AWS.ELBv2.RegisterTargets

Description

Registers the specified targets with the specified target group.

By default, the load balancer routes requests to registered targets using the protocol and port number for the target group. Alternatively, you can override the port for a target when you register it.

The target must be in the virtual private cloud (VPC) that you specified for the target group. If the target is an EC2 instance, it must be in the running state when you register it.

To remove a target from a target group, use DeregisterTargets .

Creating a Request

registerTargets

Arguments

:: Text
rtTargetGroupARN

-> RegisterTargets

Creates a value of RegisterTargets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtTargetGroupARN - The Amazon Resource Name (ARN) of the target group.

rtTargets - The targets. The default port for a target is the port for the target group. You can specify a port override. If a target is already registered, you can register it again using a different port.

data RegisterTargets

See: registerTargets smart constructor.

Request Lenses

rtTargetGroupARN :: Lens' RegisterTargets Text

The Amazon Resource Name (ARN) of the target group.

rtTargets :: Lens' RegisterTargets [TargetDescription]

The targets. The default port for a target is the port for the target group. You can specify a port override. If a target is already registered, you can register it again using a different port.

Destructuring the Response

registerTargetsResponse

Arguments

:: Int
rrsResponseStatus

-> RegisterTargetsResponse

Creates a value of RegisterTargetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrsResponseStatus - -- | The response status code.

data RegisterTargetsResponse

See: registerTargetsResponse smart constructor.

Response Lenses

rrsResponseStatus :: Lens' RegisterTargetsResponse Int

- | The response status code.