Network.AWS.ELBv2.SetIPAddressType

Description

Sets the type of IP addresses used by the subnets of the specified Application Load Balancer.

Creating a Request

setIPAddressType

Arguments

:: Text
siatLoadBalancerARN

-> IPAddressType
siatIPAddressType

-> SetIPAddressType

Creates a value of SetIPAddressType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

siatLoadBalancerARN - The Amazon Resource Name (ARN) of the load balancer.

siatIPAddressType - The IP address type. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4 .

data SetIPAddressType

See: setIPAddressType smart constructor.

Request Lenses

siatLoadBalancerARN :: Lens' SetIPAddressType Text

The Amazon Resource Name (ARN) of the load balancer.

siatIPAddressType :: Lens' SetIPAddressType IPAddressType

The IP address type. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4 .

Destructuring the Response

setIPAddressTypeResponse

Arguments

:: Int
siatrsResponseStatus

-> SetIPAddressTypeResponse

Creates a value of SetIPAddressTypeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

siatrsIPAddressType - The IP address type.

siatrsResponseStatus - -- | The response status code.

data SetIPAddressTypeResponse

See: setIPAddressTypeResponse smart constructor.

Response Lenses

siatrsIPAddressType :: Lens' SetIPAddressTypeResponse (Maybe IPAddressType)

The IP address type.

siatrsResponseStatus :: Lens' SetIPAddressTypeResponse Int

- | The response status code.