Network.AWS.ELBv2.SetRulePriorities

Description

Sets the priorities of the specified rules.

You can reorder the rules as long as there are no priority conflicts in the new order. Any existing rules that you do not specify retain their current priority.

Creating a Request

setRulePriorities :: SetRulePriorities

Creates a value of SetRulePriorities with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srpRulePriorities - The rule priorities.

data SetRulePriorities

See: setRulePriorities smart constructor.

Request Lenses

srpRulePriorities :: Lens' SetRulePriorities [RulePriorityPair]

The rule priorities.

Destructuring the Response

setRulePrioritiesResponse

Arguments

:: Int
srprsResponseStatus

-> SetRulePrioritiesResponse

Creates a value of SetRulePrioritiesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srprsRules - Information about the rules.

srprsResponseStatus - -- | The response status code.

data SetRulePrioritiesResponse

See: setRulePrioritiesResponse smart constructor.

Response Lenses

srprsRules :: Lens' SetRulePrioritiesResponse [Rule]

Information about the rules.

srprsResponseStatus :: Lens' SetRulePrioritiesResponse Int

- | The response status code.