Network.AWS.ELBv2.SetSubnets

Description

Enables the Availability Zone for the specified subnets for the specified load balancer. The specified subnets replace the previously enabled subnets.

Creating a Request

setSubnets

Arguments

:: Text
ssLoadBalancerARN

-> SetSubnets

Creates a value of SetSubnets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssLoadBalancerARN - The Amazon Resource Name (ARN) of the load balancer.

ssSubnets - The IDs of the subnets. You must specify at least two subnets. You can add only one subnet per Availability Zone.

data SetSubnets

See: setSubnets smart constructor.

Request Lenses

ssLoadBalancerARN :: Lens' SetSubnets Text

The Amazon Resource Name (ARN) of the load balancer.

ssSubnets :: Lens' SetSubnets [Text]

The IDs of the subnets. You must specify at least two subnets. You can add only one subnet per Availability Zone.

Destructuring the Response

setSubnetsResponse

Arguments

:: Int
ssrsResponseStatus

-> SetSubnetsResponse

Creates a value of SetSubnetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssrsAvailabilityZones - Information about the subnet and Availability Zone.

ssrsResponseStatus - -- | The response status code.

data SetSubnetsResponse

See: setSubnetsResponse smart constructor.

Response Lenses

ssrsAvailabilityZones :: Lens' SetSubnetsResponse [AvailabilityZone]

Information about the subnet and Availability Zone.

ssrsResponseStatus :: Lens' SetSubnetsResponse Int

- | The response status code.