Network.AWS.ELBv2.Types

Description

Service Configuration

eLBv2 :: Service

API version 2015-12-01 of the Amazon Elastic Load Balancing SDK configuration.

Errors

_InvalidConfigurationRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The requested configuration is not valid.

_SubnetNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet does not exist.

_TooManyTargetsException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of targets.

_RuleNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified rule does not exist.

_InvalidSubnetException :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet is out of available addresses.

_TooManyRulesException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of rules per load balancer.

_TooManyTargetGroupsException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of target groups for your AWS account.

_DuplicateLoadBalancerNameException :: AsError a => Getting (First ServiceError) a ServiceError

A load balancer with the specified name already exists.

_IncompatibleProtocolsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified configuration is not valid with this protocol.

_TooManyCertificatesException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of certificates per listener.

_DuplicateTagKeysException :: AsError a => Getting (First ServiceError) a ServiceError

A tag key was specified more than once.

_DuplicateListenerException :: AsError a => Getting (First ServiceError) a ServiceError

A listener with the specified port already exists.

_TooManyTagsException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of tags per load balancer.

_DuplicateTargetGroupNameException :: AsError a => Getting (First ServiceError) a ServiceError

A target group with the specified name already exists.

_HealthUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The health of the specified targets could not be retrieved due to an internal error.

_PriorityInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified priority is in use.

_TooManyLoadBalancersException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of load balancers for your AWS account.

_UnsupportedProtocolException :: AsError a => Getting (First ServiceError) a ServiceError

The specified protocol is not supported.

_InvalidTargetException :: AsError a => Getting (First ServiceError) a ServiceError

The specified target does not exist or is not in the same VPC as the target group.

_InvalidSecurityGroupException :: AsError a => Getting (First ServiceError) a ServiceError

The specified security group does not exist.

_TargetGroupNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified target group does not exist.

_ListenerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified listener does not exist.

_TooManyRegistrationsForTargetIdException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of times a target can be registered with a load balancer.

_TooManyListenersException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of listeners per load balancer.

_TargetGroupAssociationLimitException :: AsError a => Getting (First ServiceError) a ServiceError

You've reached the limit on the number of load balancers per target group.

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

This operation is not allowed.

_SSLPolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified SSL policy does not exist.

_InvalidSchemeException :: AsError a => Getting (First ServiceError) a ServiceError

The requested scheme is not valid.

_LoadBalancerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified load balancer does not exist.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

A specified resource is in use.

_CertificateNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified certificate does not exist.

ActionTypeEnum

data ActionTypeEnum

Constructors

Forward

IPAddressType

data IPAddressType

Constructors

Dualstack

IPV4

LoadBalancerSchemeEnum

data LoadBalancerSchemeEnum

Constructors

Internal

InternetFacing

LoadBalancerStateEnum

data LoadBalancerStateEnum

Constructors

Active

Failed

Provisioning

LoadBalancerTypeEnum

data LoadBalancerTypeEnum

Constructors

Application

ProtocolEnum

data ProtocolEnum

Constructors

HTTP

HTTPS

TargetHealthReasonEnum

data TargetHealthReasonEnum

Constructors

Elb_InitialHealthChecking

Elb_InternalError

Elb_RegistrationInProgress

Target_DeregistrationInProgress

Target_FailedHealthChecks

Target_InvalidState

Target_NotInUse

Target_NotRegistered

Target_ResponseCodeMismatch

Target_Timeout

TargetHealthStateEnum

data TargetHealthStateEnum

Constructors

Draining

Healthy

Initial

Unhealthy

Unused

Action

data Action

Information about an action.

See: action smart constructor.

action

Arguments

:: ActionTypeEnum
aType

-> Text
aTargetGroupARN

-> Action

Creates a value of Action with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aType - The type of action.

aTargetGroupARN - The Amazon Resource Name (ARN) of the target group.

aType :: Lens' Action ActionTypeEnum

The type of action.

aTargetGroupARN :: Lens' Action Text

The Amazon Resource Name (ARN) of the target group.

AvailabilityZone

data AvailabilityZone

Information about an Availability Zone.

See: availabilityZone smart constructor.

availabilityZone :: AvailabilityZone

Creates a value of AvailabilityZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

azSubnetId - The ID of the subnet.

azZoneName - The name of the Availability Zone.

azSubnetId :: Lens' AvailabilityZone (Maybe Text)

The ID of the subnet.

azZoneName :: Lens' AvailabilityZone (Maybe Text)

The name of the Availability Zone.

Certificate

data Certificate

Information about an SSL server certificate deployed on a load balancer.

See: certificate smart constructor.

certificate :: Certificate

Creates a value of Certificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cCertificateARN - The Amazon Resource Name (ARN) of the certificate.

cCertificateARN :: Lens' Certificate (Maybe Text)

The Amazon Resource Name (ARN) of the certificate.

Cipher

data Cipher

Information about a cipher used in a policy.

See: cipher smart constructor.

cipher :: Cipher

Creates a value of Cipher with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cPriority - The priority of the cipher.

cName - The name of the cipher.

cPriority :: Lens' Cipher (Maybe Int)

The priority of the cipher.

cName :: Lens' Cipher (Maybe Text)

The name of the cipher.

Limit

data Limit

Information about an Elastic Load Balancing resource limit for your AWS account.

See: limit smart constructor.

limit :: Limit

Creates a value of Limit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lMax - The maximum value of the limit.

lName - The name of the limit. The possible values are: * application-load-balancers * listeners-per-application-load-balancer * rules-per-application-load-balancer * target-groups * targets-per-application-load-balancer

lMax :: Lens' Limit (Maybe Text)

The maximum value of the limit.

lName :: Lens' Limit (Maybe Text)

The name of the limit. The possible values are: * application-load-balancers * listeners-per-application-load-balancer * rules-per-application-load-balancer * target-groups * targets-per-application-load-balancer

Listener

data Listener

Information about a listener.

See: listener smart constructor.

listener :: Listener

Creates a value of Listener with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lSSLPolicy - The security policy that defines which ciphers and protocols are supported. The default is the current predefined security policy.

lListenerARN - The Amazon Resource Name (ARN) of the listener.

lProtocol - The protocol for connections from clients to the load balancer.

lDefaultActions - The default actions for the listener.

lCertificates - The SSL server certificate. You must provide a certificate if the protocol is HTTPS.

lLoadBalancerARN - The Amazon Resource Name (ARN) of the load balancer.

lPort - The port on which the load balancer is listening.

lSSLPolicy :: Lens' Listener (Maybe Text)

The security policy that defines which ciphers and protocols are supported. The default is the current predefined security policy.

lListenerARN :: Lens' Listener (Maybe Text)

The Amazon Resource Name (ARN) of the listener.

lProtocol :: Lens' Listener (Maybe ProtocolEnum)

The protocol for connections from clients to the load balancer.

lDefaultActions :: Lens' Listener [Action]

The default actions for the listener.

lCertificates :: Lens' Listener [Certificate]

The SSL server certificate. You must provide a certificate if the protocol is HTTPS.

lLoadBalancerARN :: Lens' Listener (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

lPort :: Lens' Listener (Maybe Natural)

The port on which the load balancer is listening.

LoadBalancer

data LoadBalancer

Information about a load balancer.

See: loadBalancer smart constructor.

loadBalancer :: LoadBalancer

Creates a value of LoadBalancer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbState - The state of the load balancer.

lbSecurityGroups - The IDs of the security groups for the load balancer.

lbLoadBalancerName - The name of the load balancer.

lbCreatedTime - The date and time the load balancer was created.

lbVPCId - The ID of the VPC for the load balancer.

lbCanonicalHostedZoneId - The ID of the Amazon Route 53 hosted zone associated with the load balancer.

lbAvailabilityZones - The Availability Zones for the load balancer.

lbLoadBalancerARN - The Amazon Resource Name (ARN) of the load balancer.

lbIPAddressType - The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

lbScheme - The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the Internet. The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can only route requests from clients with access to the VPC for the load balancer.

lbType - The type of load balancer.

lbDNSName - The public DNS name of the load balancer.

lbState :: Lens' LoadBalancer (Maybe LoadBalancerState)

The state of the load balancer.

lbSecurityGroups :: Lens' LoadBalancer [Text]

The IDs of the security groups for the load balancer.

lbLoadBalancerName :: Lens' LoadBalancer (Maybe Text)

The name of the load balancer.

lbCreatedTime :: Lens' LoadBalancer (Maybe UTCTime)

The date and time the load balancer was created.

lbVPCId :: Lens' LoadBalancer (Maybe Text)

The ID of the VPC for the load balancer.

lbCanonicalHostedZoneId :: Lens' LoadBalancer (Maybe Text)

The ID of the Amazon Route 53 hosted zone associated with the load balancer.

lbAvailabilityZones :: Lens' LoadBalancer [AvailabilityZone]

The Availability Zones for the load balancer.

lbLoadBalancerARN :: Lens' LoadBalancer (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

lbIPAddressType :: Lens' LoadBalancer (Maybe IPAddressType)

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses).

lbScheme :: Lens' LoadBalancer (Maybe LoadBalancerSchemeEnum)

The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the Internet. The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can only route requests from clients with access to the VPC for the load balancer.

lbType :: Lens' LoadBalancer (Maybe LoadBalancerTypeEnum)

The type of load balancer.

lbDNSName :: Lens' LoadBalancer (Maybe Text)

The public DNS name of the load balancer.

LoadBalancerAttribute

data LoadBalancerAttribute

Information about a load balancer attribute.

See: loadBalancerAttribute smart constructor.

loadBalancerAttribute :: LoadBalancerAttribute

Creates a value of LoadBalancerAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbaValue - The value of the attribute.

lbaKey - The name of the attribute. * access_logs.s3.enabled - Indicates whether access logs stored in Amazon S3 are enabled. The value is true or false . * access_logs.s3.bucket - The name of the S3 bucket for the access logs. This attribute is required if access logs in Amazon S3 are enabled. The bucket must exist in the same region as the load balancer and have a bucket policy that grants Elastic Load Balancing permission to write to the bucket. * access_logs.s3.prefix - The prefix for the location in the S3 bucket. If you don't specify a prefix, the access logs are stored in the root of the bucket. * deletion_protection.enabled - Indicates whether deletion protection is enabled. The value is true or false . * idle_timeout.timeout_seconds - The idle timeout value, in seconds. The valid range is 1-3600. The default is 60 seconds.

lbaValue :: Lens' LoadBalancerAttribute (Maybe Text)

The value of the attribute.

lbaKey :: Lens' LoadBalancerAttribute (Maybe Text)

The name of the attribute. * access_logs.s3.enabled - Indicates whether access logs stored in Amazon S3 are enabled. The value is true or false . * access_logs.s3.bucket - The name of the S3 bucket for the access logs. This attribute is required if access logs in Amazon S3 are enabled. The bucket must exist in the same region as the load balancer and have a bucket policy that grants Elastic Load Balancing permission to write to the bucket. * access_logs.s3.prefix - The prefix for the location in the S3 bucket. If you don't specify a prefix, the access logs are stored in the root of the bucket. * deletion_protection.enabled - Indicates whether deletion protection is enabled. The value is true or false . * idle_timeout.timeout_seconds - The idle timeout value, in seconds. The valid range is 1-3600. The default is 60 seconds.

LoadBalancerState

data LoadBalancerState

Information about the state of the load balancer.

See: loadBalancerState smart constructor.

loadBalancerState :: LoadBalancerState

Creates a value of LoadBalancerState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbsReason - A description of the state.

lbsCode - The state code. The initial state of the load balancer is provisioning . After the load balancer is fully set up and ready to route traffic, its state is active . If the load balancer could not be set up, its state is failed .

lbsReason :: Lens' LoadBalancerState (Maybe Text)

A description of the state.

lbsCode :: Lens' LoadBalancerState (Maybe LoadBalancerStateEnum)

The state code. The initial state of the load balancer is provisioning . After the load balancer is fully set up and ready to route traffic, its state is active . If the load balancer could not be set up, its state is failed .

Matcher

data Matcher

Information to use when checking for a successful response from a target.

See: matcher smart constructor.

matcher

Arguments

:: Text
mHTTPCode

-> Matcher

Creates a value of Matcher with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mHTTPCode - The HTTP codes. You can specify values between 200 and 499. The default value is 200. You can specify multiple values (for example, "200,202") or a range of values (for example, "200-299").

mHTTPCode :: Lens' Matcher Text

The HTTP codes. You can specify values between 200 and 499. The default value is 200. You can specify multiple values (for example, "200,202") or a range of values (for example, "200-299").

Rule

data Rule

Information about a rule.

See: rule smart constructor.

rule :: Rule

Creates a value of Rule with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rPriority - The priority.

rActions - The actions.

rConditions - The conditions.

rRuleARN - The Amazon Resource Name (ARN) of the rule.

rIsDefault - Indicates whether this is the default rule.

rPriority :: Lens' Rule (Maybe Text)

The priority.

rActions :: Lens' Rule [Action]

The actions.

rConditions :: Lens' Rule [RuleCondition]

The conditions.

rRuleARN :: Lens' Rule (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

rIsDefault :: Lens' Rule (Maybe Bool)

Indicates whether this is the default rule.

RuleCondition

data RuleCondition

Information about a condition for a rule.

See: ruleCondition smart constructor.

ruleCondition :: RuleCondition

Creates a value of RuleCondition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcField - The name of the field. The possible values are host-header and path-pattern .

rcValues - The condition value. If the field name is host-header , you can specify a single host name (for example, my.example.com). A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * - . * * (matches 0 or more characters) * ? (matches exactly 1 character) If the field name is path-pattern , you can specify a single path pattern (for example, img*). A path pattern is case sensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * _ - . $ / ~ " ' @ : + * & (using &amp;) * * (matches 0 or more characters) * ? (matches exactly 1 character)

rcField :: Lens' RuleCondition (Maybe Text)

The name of the field. The possible values are host-header and path-pattern .

rcValues :: Lens' RuleCondition [Text]

The condition value. If the field name is host-header , you can specify a single host name (for example, my.example.com). A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * - . * * (matches 0 or more characters) * ? (matches exactly 1 character) If the field name is path-pattern , you can specify a single path pattern (for example, img*). A path pattern is case sensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters. * A-Z, a-z, 0-9 * _ - . $ / ~ " ' @ : + * & (using &amp;) * * (matches 0 or more characters) * ? (matches exactly 1 character)

RulePriorityPair

data RulePriorityPair

Information about the priorities for the rules for a listener.

See: rulePriorityPair smart constructor.

rulePriorityPair :: RulePriorityPair

Creates a value of RulePriorityPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rppPriority - The rule priority.

rppRuleARN - The Amazon Resource Name (ARN) of the rule.

rppPriority :: Lens' RulePriorityPair (Maybe Natural)

The rule priority.

rppRuleARN :: Lens' RulePriorityPair (Maybe Text)

The Amazon Resource Name (ARN) of the rule.

SSLPolicy

data SSLPolicy

Information about a policy used for SSL negotiation.

See: sslPolicy smart constructor.

sslPolicy :: SSLPolicy

Creates a value of SSLPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spCiphers - The ciphers.

spName - The name of the policy.

spSSLProtocols - The protocols.

spCiphers :: Lens' SSLPolicy [Cipher]

The ciphers.

spName :: Lens' SSLPolicy (Maybe Text)

The name of the policy.

spSSLProtocols :: Lens' SSLPolicy [Text]

The protocols.

Tag

data Tag

Information about a tag.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - The value of the tag.

tagKey - The key of the tag.

tagValue :: Lens' Tag (Maybe Text)

The value of the tag.

tagKey :: Lens' Tag Text

The key of the tag.

TagDescription

data TagDescription

The tags associated with a resource.

See: tagDescription smart constructor.

tagDescription :: TagDescription

Creates a value of TagDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdResourceARN - The Amazon Resource Name (ARN) of the resource.

tdTags - Information about the tags.

tdResourceARN :: Lens' TagDescription (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

tdTags :: Lens' TagDescription (Maybe (NonEmpty Tag))

Information about the tags.

TargetDescription

data TargetDescription

Information about a target.

See: targetDescription smart constructor.

targetDescription

Arguments

:: Text
tdId

-> TargetDescription

Creates a value of TargetDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdPort - The port on which the target is listening.

tdId - The ID of the target.

tdPort :: Lens' TargetDescription (Maybe Natural)

The port on which the target is listening.

tdId :: Lens' TargetDescription Text

The ID of the target.

TargetGroup

data TargetGroup

Information about a target group.

See: targetGroup smart constructor.

targetGroup :: TargetGroup

Creates a value of TargetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tgMatcher - The HTTP codes to use when checking for a successful response from a target.

tgHealthCheckPath - The destination for the health check request.

tgUnhealthyThresholdCount - The number of consecutive health check failures required before considering the target unhealthy.

tgVPCId - The ID of the VPC for the targets.

tgTargetGroupARN - The Amazon Resource Name (ARN) of the target group.

tgProtocol - The protocol to use for routing traffic to the targets.

tgHealthCheckIntervalSeconds - The approximate amount of time, in seconds, between health checks of an individual target.

tgHealthyThresholdCount - The number of consecutive health checks successes required before considering an unhealthy target healthy.

tgHealthCheckProtocol - The protocol to use to connect with the target.

tgLoadBalancerARNs - The Amazon Resource Names (ARN) of the load balancers that route traffic to this target group.

tgHealthCheckTimeoutSeconds - The amount of time, in seconds, during which no response means a failed health check.

tgHealthCheckPort - The port to use to connect with the target.

tgTargetGroupName - The name of the target group.

tgPort - The port on which the targets are listening.

tgMatcher :: Lens' TargetGroup (Maybe Matcher)

The HTTP codes to use when checking for a successful response from a target.

tgHealthCheckPath :: Lens' TargetGroup (Maybe Text)

The destination for the health check request.

tgUnhealthyThresholdCount :: Lens' TargetGroup (Maybe Natural)

The number of consecutive health check failures required before considering the target unhealthy.

tgVPCId :: Lens' TargetGroup (Maybe Text)

The ID of the VPC for the targets.

tgTargetGroupARN :: Lens' TargetGroup (Maybe Text)

The Amazon Resource Name (ARN) of the target group.

tgProtocol :: Lens' TargetGroup (Maybe ProtocolEnum)

The protocol to use for routing traffic to the targets.

tgHealthCheckIntervalSeconds :: Lens' TargetGroup (Maybe Natural)

The approximate amount of time, in seconds, between health checks of an individual target.

tgHealthyThresholdCount :: Lens' TargetGroup (Maybe Natural)

The number of consecutive health checks successes required before considering an unhealthy target healthy.

tgHealthCheckProtocol :: Lens' TargetGroup (Maybe ProtocolEnum)

The protocol to use to connect with the target.

tgLoadBalancerARNs :: Lens' TargetGroup [Text]

The Amazon Resource Names (ARN) of the load balancers that route traffic to this target group.

tgHealthCheckTimeoutSeconds :: Lens' TargetGroup (Maybe Natural)

The amount of time, in seconds, during which no response means a failed health check.

tgHealthCheckPort :: Lens' TargetGroup (Maybe Text)

The port to use to connect with the target.

tgTargetGroupName :: Lens' TargetGroup (Maybe Text)

The name of the target group.

tgPort :: Lens' TargetGroup (Maybe Natural)

The port on which the targets are listening.

TargetGroupAttribute

data TargetGroupAttribute

Information about a target group attribute.

See: targetGroupAttribute smart constructor.

targetGroupAttribute :: TargetGroupAttribute

Creates a value of TargetGroupAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tgaValue - The value of the attribute.

tgaKey - The name of the attribute. * deregistration_delay.timeout_seconds - The amount time for Elastic Load Balancing to wait before changing the state of a deregistering target from draining to unused . The range is 0-3600 seconds. The default value is 300 seconds. * stickiness.enabled - Indicates whether sticky sessions are enabled. The value is true or false . * stickiness.type - The type of sticky sessions. The possible value is lb_cookie . * stickiness.lb_cookie.duration_seconds - The time period, in seconds, during which requests from a client should be routed to the same target. After this time period expires, the load balancer-generated cookie is considered stale. The range is 1 second to 1 week (604800 seconds). The default value is 1 day (86400 seconds).

tgaValue :: Lens' TargetGroupAttribute (Maybe Text)

The value of the attribute.

tgaKey :: Lens' TargetGroupAttribute (Maybe Text)

The name of the attribute. * deregistration_delay.timeout_seconds - The amount time for Elastic Load Balancing to wait before changing the state of a deregistering target from draining to unused . The range is 0-3600 seconds. The default value is 300 seconds. * stickiness.enabled - Indicates whether sticky sessions are enabled. The value is true or false . * stickiness.type - The type of sticky sessions. The possible value is lb_cookie . * stickiness.lb_cookie.duration_seconds - The time period, in seconds, during which requests from a client should be routed to the same target. After this time period expires, the load balancer-generated cookie is considered stale. The range is 1 second to 1 week (604800 seconds). The default value is 1 day (86400 seconds).

TargetHealth

data TargetHealth

Information about the current health of a target.

See: targetHealth smart constructor.

targetHealth :: TargetHealth

Creates a value of TargetHealth with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

thState - The state of the target.

thReason - The reason code. If the target state is healthy , a reason code is not provided. If the target state is initial , the reason code can be one of the following values: * Elb.RegistrationInProgress - The target is in the process of being registered with the load balancer. * Elb.InitialHealthChecking - The load balancer is still sending the target the minimum number of health checks required to determine its health status. If the target state is unhealthy , the reason code can be one of the following values: * Target.ResponseCodeMismatch - The health checks did not return an expected HTTP code. * Target.Timeout - The health check requests timed out. * Target.FailedHealthChecks - The health checks failed because the connection to the target timed out, the target response was malformed, or the target failed the health check for an unknown reason. * Elb.InternalError - The health checks failed due to an internal error. If the target state is unused , the reason code can be one of the following values: * Target.NotRegistered - The target is not registered with the target group. * Target.NotInUse - The target group is not used by any load balancer or the target is in an Availability Zone that is not enabled for its load balancer. * Target.InvalidState - The target is in the stopped or terminated state. If the target state is draining , the reason code can be the following value: * Target.DeregistrationInProgress - The target is in the process of being deregistered and the deregistration delay period has not expired.

thDescription - A description of the target health that provides additional details. If the state is healthy , a description is not provided.

thState :: Lens' TargetHealth (Maybe TargetHealthStateEnum)

The state of the target.

thReason :: Lens' TargetHealth (Maybe TargetHealthReasonEnum)

The reason code. If the target state is healthy , a reason code is not provided. If the target state is initial , the reason code can be one of the following values: * Elb.RegistrationInProgress - The target is in the process of being registered with the load balancer. * Elb.InitialHealthChecking - The load balancer is still sending the target the minimum number of health checks required to determine its health status. If the target state is unhealthy , the reason code can be one of the following values: * Target.ResponseCodeMismatch - The health checks did not return an expected HTTP code. * Target.Timeout - The health check requests timed out. * Target.FailedHealthChecks - The health checks failed because the connection to the target timed out, the target response was malformed, or the target failed the health check for an unknown reason. * Elb.InternalError - The health checks failed due to an internal error. If the target state is unused , the reason code can be one of the following values: * Target.NotRegistered - The target is not registered with the target group. * Target.NotInUse - The target group is not used by any load balancer or the target is in an Availability Zone that is not enabled for its load balancer. * Target.InvalidState - The target is in the stopped or terminated state. If the target state is draining , the reason code can be the following value: * Target.DeregistrationInProgress - The target is in the process of being deregistered and the deregistration delay period has not expired.

thDescription :: Lens' TargetHealth (Maybe Text)

A description of the target health that provides additional details. If the state is healthy , a description is not provided.

TargetHealthDescription

data TargetHealthDescription

Information about the health of a target.

See: targetHealthDescription smart constructor.

targetHealthDescription :: TargetHealthDescription

Creates a value of TargetHealthDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

thdTargetHealth - The health information for the target.

thdHealthCheckPort - The port to use to connect with the target.

thdTarget - The description of the target.

thdTargetHealth :: Lens' TargetHealthDescription (Maybe TargetHealth)

The health information for the target.

thdHealthCheckPort :: Lens' TargetHealthDescription (Maybe Text)

The port to use to connect with the target.

thdTarget :: Lens' TargetHealthDescription (Maybe TargetDescription)

The description of the target.